refactor(bin): update data collection script
Turns out, there is not benefit to storing the `source` column as a list since pandas.read_csv converts this to a string and then we have to use ast.literaleval to convert back to list.

Since we ast.parse() expects a string as input anyway, I have modified bin/data-collection.py such that the `source` column is directly saved as a string.